Highlight : for main:
As per @wcrissman’s comments, we want to highlight the colon for main
as well to be consistent. I discovered this caused a precedence issue
with the highlighting so I had to reorder somethings as well to get it
right. Perhaps the problem would really be solved better by regions
expressions, but this works fine for now.